requirements.txt.txt
google-cloud-pubsub
mysql-connector-python